Scroll-offest temporary implementation, appending animation and errorText colour
pokeContainer will scroll to the bottom automatically after adding a Pokémon so the user can see it. However, the newly added card was being cut-off due to scrollHeight not moving to bottom, being offset. A temporary fix for this has been to add padding to the bottom of the container, so the card is in full view. More research to be done into scrollHeight and how to fix this problem properly.
New Pokémon are animated into view upon appending.
errorText now correctly displays red on error and green on success.