Template CSS and removed comments
Commented code will be available to see in previous commits, so there's no need to keep it lurking in the code and cluttering it. It's been removed.
There was a "flashing" effect on the first render of a Pokémon card as the card would load in, then have styles applied to it, then scale from 0 to 1. This all proved to be quite... well, crap. Once the first card rendered, the external CSS had been cached and then the succeeding cards were loading as normal.
By moving the styles from the external CSS into inline styles within the template, that flashing is now gone and all cards render as expected.